requirements.txt
fastapi
uvicorn
python-multipart
spotipy
transformers
torch
pillow
python-dotenv
